+++ operations.go
@@ -1,11 +1,15 @@
 package repeat
 
+// Operation is the type of function for repetition.
+type Operation func(error) error
+
 // LimitMaxTries returns true if attempt number is less then max.
 func LimitMaxTries(max int) Operation {
 	return FnWithErrorAndCounter(func(e error, c int) error {
 		if c < max {
 			return e
 		}
+
 		return &StopError{e}
 	})
 }
@@ -16,7 +20,35 @@ func StopOnSuccess() Operation {
 		if e != nil {
 			return e
 		}
+
 		return &StopError{e}
+	}
+}
+
+// FnOnSuccess executes operation in case of error is nil.
+func FnOnSuccess(op func(error) error) Operation {
+	return func(e error) error {
+		if e != nil {
+			return e
+		}
+
+		return op(e)
+	}
+}
+
+// FnHintTemporary hints all operation errors as temporary.
+func FnHintTemporary(op func(error) error) Operation {
+	return func(e error) error {
+		err := op(e)
+		switch err.(type) {
+		case nil:
+		case *TemporaryError:
+		case *StopError:
+		default:
+			err = HintTemporary(err)
+		}
+
+		return err
 	}
 }
 
@@ -40,5 +72,21 @@ func FnWithCounter(op func(int) error) Operation {
 func Fn(op func() error) Operation {
 	return func(_ error) error {
 		return op()
+	}
+}
+
+// FnS wrap operation with no arguments and return value.
+func FnS(op func()) Operation {
+	return func(e error) error {
+		op()
+		return e
+	}
+}
+
+// FnES wrap operation with no return value.
+func FnES(op func(error)) Operation {
+	return func(e error) error {
+		op(e)
+		return e
 	}
 }