+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 # 0.1.0
 
 * initial version, all tests passed
+
+# 1.0.0
+
+* public release
+* add new interface Repeater
+* add new concept - OpWrapper
+* add new function to create Repeater with wrapping operations
+* add new WithContext - a repeater that checks for context errors before operation call